+++ tools/compilers/aas/AASFile.cpp
@@ -1195,7 +1195,7 @@ idAASFileLocal::MemorySize
 ================
 */
 int idAASFileLocal::MemorySize( void ) const {
-	int size;
+	size_t size;
 
 	size = planeList.Size();
 	size += vertices.Size();
@@ -1210,7 +1210,7 @@ int idAASFileLocal::MemorySize( void ) const {
 	size += clusters.Size();
 	size += sizeof( idReachability_Walk ) * NumReachabilities();
 
-	return size;
+	return static_cast<int>(size);
 }
 
 /*
