+++ tools/materialeditor/MaterialDoc.cpp
@@ -75,7 +75,7 @@ void MaterialDoc::SetRenderMaterial(idMaterial* material, bool parseMaterial, bo
 	material->GetText( declText );
 
 	renderMaterial->GetText(declText);
-	src.LoadMemory(declText, strlen(declText), "Material");
+    src.LoadMemory(declText, static_cast<int>(strlen(declText)), "Material");
 
 	ParseMaterial(&src);
 }
@@ -655,7 +655,7 @@ void MaterialDoc::ParseMaterialText(const char* source) {
 	src.SkipUntilString("{");*/
 
 	//Now let the material parse the text
-	renderMaterial->Parse(source, strlen(source));
+    renderMaterial->Parse(source, static_cast<int>(strlen(source)));
 }
 
 /**
